requirements.txt
pyyaml
msgpack==1.0.0
pendulum
click
pandas
requests
kafka-python==2.0.2